Lock the derivepassphrase CLI against concurrent updating
In a very coarse manner, after command-line parsing, detect whether the
CLI operation is a read-write operation or a read-only one, and if it is
read-write, run the whole operation while holding a lock.

The lock is held on a temporary file whose basename is dependent on the
full path to the `derivepassphrase` configuration directory.  The file
is stored in the system temporary directory if possible, and will be
synchronized using `msvcrt.locking` or `fcntl.flock`, whichever is
appropriate.

Coverage checking for the Windows-specific application code is disabled
because we have no Windows runners for the test suite.  The test suite
itself merely adds tests for the temporary directory function, but is
otherwise unchanged.