Reimplement the `vault` CLI as methods of a context object
Introduce a "vault context" object holding the associated data for the
call to the `derivepassphrase vault` CLI, and implement the actual CLI
as a collection of methods on this object.  The "vault context" object
mostly just wraps the click context object and encodes the dispatch
table for the CLI operations.  The main advantage of this modelling is
the much smaller size per, and lower cyclomatic complexity of, each CLI
operation.

As an implementation detail, the old implementation of the
`get_user_config` inner function's error handling branch for `OSError`
had coverage testing purely by accident, because `get_user_config` was
being called unconditionally, very early in the CLI control flow.  The
refactored CLI however reads the user configuration only as needed for
the specific operation. So this error handling branch needs a new,
separate test.

As a further implementation detail, the `is_param_set` inner function no
longer caches its result, because it is now an object method:
`functools.cache` explicitly documents that adding a cache to object
methods causes a circular reference between the cache and the object via
the `self` parameter.  Given the actual contents of this function, there
appears to be little point in working around this memory issue just to
outfit an already cheap operation (a single `Mapping.get` call) with
a cache.

As yet another further implementation detail, checking for misleading
passphrases should work even if the passed configuration snippet is not
technically a `dict`, but rather a `Mapping`.

Finally, due to the vault context object now wrapping the click context,
it is possible (and sensible) to wrap the logging calls into methods of
the vault context object so that less out-of-bounds information needs to
be passed to the logger explicitly at the call site.