Wrap click.testing.CliRunner to directly return a ReadableResult
Instead of using `click.testing.CliRunner` directly and transforming the
result, wrap the runner in a new class that does the necessary
transformations on the result automatically.

This indirection of the CLI runner also allows the wrapper to cope with
(some) incompatible changes in `click` across versions, or even
eventually swap out the underlying implementation completely.

As a first incompatibility to bridge, `click` 8.2.0 (currently in beta)
merges many bad or incomplete fixes for long-standing issues, but with
no regard to backward compatibility.  Of particular relevance,
`click.testing.CliRunner` no longer accepts the `mix_stderr`
argument,[^1] and the meanings of the `output` and the `stdout`
attribute of `click.testing.Result` have changed.[^2]  These two
incompatibilities are handled by the wrapper.

  [^1]: This breaks practically every call to the runner: `mix_stderr`
  defaults to `True`, thus the default behavior is to mangle the output
  streams.  Why would you want that, in your testing machinery no less,
  where automatically mangling or discarding information behind the
  scenes is detrimental to debugging?!  As a consequence, `mix_stderr`
  *has* to be specified if you want output streams not to be mangled,
  and if you want them to be mangled, you *should* specify `mix_stderr`
  explicitly.  So basically, you always specify `mix_stderr`.  And thus
  removing the (definition of the) `mix_stderr` argument breaks
  practically every call to the runner.

  [^2]: This is complementary to the removal of the `mix_stderr`
  argument: the mixed output is now always stored in the `output`
  attribute, and the `stdout` is the true standard output.  Of course,
  this contradicts the previous use of `output` and `stdout`, where
  `output` was the true standard output, and `stdout` was a convenience
  alias via an extra layer of redirection.