Add tests for the configuration mutex
Conceptually, we need two types of tests for the configuration mutex:
the first type asserts that the program logic works for *any* kind of
(potentially fake) configuration mutex, and the second type asserts that
the configuration mutex works for *any* set of (potentially fake)
operations.  In both cases, the fake parts usually impose a specific
order of locking or running, controllable from the outside.

We modify the `ConfigurationMutex` class to shift the management of the
lock file descriptor into the lock and unlock functions themselves,
instead of cramming them into `__enter__` and `__exit__`.  (This means
that subclasses need not deal with the lock file, or in fact, with any
of the platform-specific locking machinery.)  A custom subclass in the
`tests` namespace provides a faked mutex that imposes a specific locking
order, controllable from the outside via message passing.  This subclass
is then used with a state machine to test various sets of operations
together with the faked mutex, i.e. to provide tests of the first type.
(The standard tests, executed with the actual mutex implementation,
already provide tests of the second type.)

Regrettably, this set of tests requires concurrency support.
Specifically, because the click testing machinery needs to reset
process-wide properties for each concurrent unit, and because the tests
require this to be done in a time-overlapped manner, the tests
specifically require multiprocessing support.  Care must also be taken
to ensure that coverage management works: that it is enabled in the
subprocesses whenever it is enabled in the parent (which is
a straightforward configuration issue), and that the parent process does
not accidentally remove these coverage measurements as junk files in
case the subprocesses are spawned in specifically prepared and
automatically cleaned up temporary directories (which is a manual labor
issue).