Add changes necessary to appease The Annoying OS
Add the changes that are necessary to appease The Annoying Operating
System, a.k.a. Microsoft Windows.

Some of these changes are OS-specific and (mostly) sensible, such as
error codes being slightly different, or some functionality not being
available.  Some of these changes are coincidental, such as linting
or type checking errors because unavailable symbols aren't even defined
for the type checker's benefit.  Finally, some changes are outright
stupid, such as Python's documentation on `tempfile.gettempdir` being
incomplete.

In the end, most changes boil down to checking for a range of error
codes instead of a specific one, or adding missing fixture calls, or
silencing type checker warnings that aren't encountered in practice, or
ruling out bad degenerate hypothesis test cases that aren't degenerate
or aren't functional at all on The Annoying OS.