Embed our "wishlist" branch in the documentation
The contents of `docs/wishlist/` now host a checkout of the `wishlist`
branch, providing our own bug and wish listing. The navigation includes
a new section "Wishlist", and the section includes custom CSS styling.
The "actions" partial template is also updated so that it generates
correct links to the raw file on GitHub, i.e., on the "wishlist" branch
instead of on the "master" branch.