Pin testing dependencies
Pin the versions of testing dependencies to ensure easy installability.
Specifically, pin `cryptography` on PyPy to versions that support
wheels, and pin `hypothesis` on Python 3.9 (CPython and PyPy) to
versions that correctly support the old LL(1) parser. Because
`cryptography` is an optional dependency, we introduce a new, matching
extra for the `cryptography` version pins, with an appropriate name.

As the extras are getting a bit confusing, especially with regard to one
another, we add a one-line synopsis to the beginning of each extra's
commentary.