Fix some coverage slipups
The first slipup deals with two different wrappers around `os.getenv`,
depending on whether the system support bytes-based environment
variables. We care about whether the wrapper function (whichever
implementation) is called correctly, not which wrapper function is
selected. So, rewrite the branch to put the system-specific definitions
under coverage exclusion, and the rest not.

The second slipup concerns the concurrency limit. The detection is
system- and runtime-dependent, but the error paths were previously not
excluded from coverage. Additionally, we claimed to manually support the
`PYTHON_CPU_COUNT` environment variable on Python 3.12, but contained no
matching code. So, rewrite the code to add that missing support, in
a coverage-friendly way.