Update the man page diagnostic script: settings and output format
Update the man page diagnostic script to current settings: use the
updated path to the man pages, and use metavars without the `!s`
conversion. Additionally, since the output is currently a diagnostic on
standard error, and very verbose, emit a terser shorthand notation to
standard output instead. (Also sort the files to be examined first.)