Link the commands, API classes, bugs and wishlist items in the changelog
Turn all mentions of a command or subcommand name, or of an API member,
into a link to that command's manpage or API member's API documentation.

Furthermore, if a changelog entry refers to a bug or wishlist item, link
to the respective item as well.

Finally, collect and deduplicate all Markdown references at the bottom
of the page.