Fix minor oversights in the documentation
The interal "Future ideas" page now also includes the same icons as the
changelog uses for referring to bugs and wishlist entries.

The wishlist path is slightly different, depending on whether MkDocs
builds in "online" or "offline" mode.  The CSS needs to accomodate both
paths.

The main API documentation should not attempt to refer to internals by
hyperlink.